requirements.txt
promptflow==1.17.2
promptflow-azure==1.17.2
deepeval==2.5.3
promptflow-core==1.17.2
promptflow-devkit==1.17.2
promptflow-evals==0.3.2
promptflow-tracing==1.17.2